pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.7.0",
    "google-genai>=1.26.0",
    "llama-index>=0.12.52",
    "llama-index-embeddings-google-genai>=0.2.1",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "markdown>=3.8.2",
    "pydantic-settings>=2.10.1",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.0",
    "uvicorn>=0.35.0",
    "xhtml2pdf>=0.2.17",
]
